Solution to Calculate Average Marks of Students

+++ Calculate Average Marks of Students@@67877d61dc107486ad5c2743/code.c
@@ -9,13 +9,13 @@ int main(){
     scanf("%d",&n);
     struct Student s[n];
     for(int i=0;i<n;i++){
-    scanf("%d  %s %.2f",&s[i].RollNo,&s[i].Name,&s[i].Marks);
+    scanf("%d  %s %f",&s[i].RollNo,&s[i].Name,&s[i].Marks);
     }
-    int total=0;
+    float total=0;
     for(int i=0;i<n;i++){
         total+=s[i].Marks;
     }
-    int average = total/n;
+    float average = total/n;
     printf("Average Marks: %.2f",average);
 
     return 0;
